Update the launch of local function : nor more use of thread
There is a issue when launching each local function in a separate thread. Although everything works fine in local, in the kube cluster it doesn't terminate always. May be the jvm of the kube cluster has different setting pertaining to threads and memory compare to the classic openjdk of ubuntu